Remove more implicit casts in upnp part
Remove more "implicit integer or enum conversions" as well as memset
before snprintf.

+++ upnp/src/genlib/net/http/statcodes.c
@@ -129,7 +129,7 @@ init_table( IN const char *encoded_str,
 
     for( i = 0; i < tbl_size; i++ ) {
         table[i] = s;
-        s += strlen( s ) + 1;   /* next entry */
+        s += strlen( s ) + (size_t)1;   /* next entry */
     }
 }
